Defeat speculative unroll in pauseMenuFn_80129ee0 cell-scan loop
u8 loop counter recomputes (u8)i index per iteration (clrlwi;slwi;add),
matching target's non-unrolled bdnz-free scan instead of 8x unroll.

+++ src/main/dll/dll_0000_gameui.c
@@ -2786,7 +2786,7 @@ void pauseMenuFn_80129ee0(void)
         }
         else
         {
-            int i;
+            u8 i;
             for (i = 0; i < 0x2d; i++)
             {
                 if (cell == tbl->cellMap[i].cell)
@@ -2794,7 +2794,7 @@ void pauseMenuFn_80129ee0(void)
                     break;
                 }
             }
-            if ((u8)i != 0x2d)
+            if (i != 0x2d)
             {
                 int code = tbl->cellMap[i].code;
                 lbl_803DD8E0 = code;
